requirements.txt
load-dotenv==0.1.0
mistralai~=0.1.2
panel~=1.3.8
